Extract ShiftedRead
